requirements.txt
Django>=5.2
djangorestframework>=3.16
drf_spectacular>=0.29
